Add JacksonConfig to enable Spring Data web support
Created a new configuration class, JacksonConfig, to configure Spring Data web support. This class enables pagination serialization via DTO, enhancing the handling of paged data responses. This will fix the warning:

Serializing PageImpl instances as-is is not supported, meaning that there is no guarantee about the stability of the resulting JSON structure!
	For a stable JSON structure, please use Spring Data's PagedModel (globally via @EnableSpringDataWebSupport(pageSerializationMode = VIA_DTO))
	or Spring HATEOAS and Spring Data's PagedResourcesAssembler as documented in https://docs.spring.io/spring-data/commons/reference/repositories/core-extensions.html#core.web.pageables.